api/requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
async-timeout==4.0.2
click==8.1.3
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
httptools==0.5.0
idna==3.4
packaging==21.3
pydantic==1.10.2
pyparsing==3.0.9
python-dotenv==0.21.0
PyYAML==6.0
redis==4.3.5
sniffio==1.3.0
SQLAlchemy==2.0.0b3
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
python-jose[cryptography]
passlib[bcrypt]